不具合修正

+++ agent.py
@@ -27,52 +27,52 @@ class Agent:
     self.load()
 
   def configure_optimizers(self, config):
-      """
-      This long function is unfortunately doing something very simple and is being very defensive:
-      We are separating out all parameters of the model into two buckets: those that will experience
-      weight decay for regularization and those that won't (biases, and layernorm/embedding weights).
-      We are then returning the PyTorch optimizer object.
-      """
+    """
+    This long function is unfortunately doing something very simple and is being very defensive:
+    We are separating out all parameters of the model into two buckets: those that will experience
+    weight decay for regularization and those that won't (biases, and layernorm/embedding weights).
+    We are then returning the PyTorch optimizer object.
+    """
 
-      # separate out all parameters to those that will and won't experience regularizing weight decay
-      decay = set()
-      no_decay = set()
-      # whitelist_weight_modules = (torch.nn.Linear, )
-      whitelist_weight_modules = (torch.nn.Linear, torch.nn.Conv2d)
-      blacklist_weight_modules = (torch.nn.LayerNorm, torch.nn.Embedding)
-      for mn, m in self.net.named_modules():
-          for pn, p in m.named_parameters():
-              fpn = '%s.%s' % (mn, pn) if mn else pn # full param name
+    # separate out all parameters to those that will and won't experience regularizing weight decay
+    decay = set()
+    no_decay = set()
+    # whitelist_weight_modules = (torch.nn.Linear, )
+    whitelist_weight_modules = (torch.nn.Linear, torch.nn.Conv2d)
+    blacklist_weight_modules = (torch.nn.LayerNorm, torch.nn.Embedding)
+    for mn, m in self.net.named_modules():
+      for pn, p in m.named_parameters():
+        fpn = '%s.%s' % (mn, pn) if mn else pn  # full param name
 
-              if pn.endswith('bias'):
-                  # all biases will not be decayed
-                  no_decay.add(fpn)
-              elif pn.endswith('weight') and isinstance(m, whitelist_weight_modules):
-                  # weights of whitelist modules will be weight decayed
-                  decay.add(fpn)
-              elif pn.endswith('weight') and isinstance(m, blacklist_weight_modules):
-                  # weights of blacklist modules will NOT be weight decayed
-                  no_decay.add(fpn)
+        if pn.endswith('bias'):
+          # all biases will not be decayed
+          no_decay.add(fpn)
+        elif pn.endswith('weight') and isinstance(m, whitelist_weight_modules):
+          # weights of whitelist modules will be weight decayed
+          decay.add(fpn)
+        elif pn.endswith('weight') and isinstance(m, blacklist_weight_modules):
+          # weights of blacklist modules will NOT be weight decayed
+          no_decay.add(fpn)
 
-      # special case the position embedding parameter in the root GPT module as not decayed
-      no_decay.add('pos_emb')
-      no_decay.add('global_pos_emb')
+    # special case the position embedding parameter in the root GPT module as not decayed
+    no_decay.add('pos_emb')
+    no_decay.add('global_pos_emb')
 
-      # validate that we considered every parameter
-      param_dict = {pn: p for pn, p in self.net.named_parameters()}
-      inter_params = decay & no_decay
-      union_params = decay | no_decay
-      assert len(inter_params) == 0, "parameters %s made it into both decay/no_decay sets!" % (str(inter_params), )
-      assert len(param_dict.keys() - union_params) == 0, "parameters %s were not separated into either decay/no_decay set!" \
-                                                  % (str(param_dict.keys() - union_params), )
+    # validate that we considered every parameter
+    param_dict = {pn: p for pn, p in self.net.named_parameters()}
+    inter_params = decay & no_decay
+    union_params = decay | no_decay
+    assert len(inter_params) == 0, "parameters %s made it into both decay/no_decay sets!" % (str(inter_params), )
+    assert len(param_dict.keys() - union_params) == 0, "parameters %s were not separated into either decay/no_decay set!" \
+        % (str(param_dict.keys() - union_params), )
 
-      # create the pytorch optimizer object
-      optim_groups = [
-          {"params": [param_dict[pn] for pn in sorted(list(decay))], "weight_decay": config.weight_decay},
-          {"params": [param_dict[pn] for pn in sorted(list(no_decay))], "weight_decay": 0.0},
-      ]
-      optimizer = torch.optim.AdamW(optim_groups, lr=config.adam_lr, betas=config.adam_beta)
-      return optimizer
+    # create the pytorch optimizer object
+    optim_groups = [
+        {"params": [param_dict[pn] for pn in sorted(list(decay))], "weight_decay": config.weight_decay},
+        {"params": [param_dict[pn] for pn in sorted(list(no_decay))], "weight_decay": 0.0},
+    ]
+    optimizer = torch.optim.AdamW(optim_groups, lr=config.adam_lr, betas=config.adam_beta)
+    return optimizer
 
   def load(self):
     if os.path.exists(self._config.checkpoint_path):
@@ -132,7 +132,7 @@ class Agent:
       input = Input(
         rtg=torch.tensor(np.array(rtgs)[-self._config.context_length:]).reshape(1, -1, 1).to(torch.float32).to(self.device),
         state=torch.tensor(np.array(states)[-self._config.context_length:]).unsqueeze(0).to(torch.float32).to(self.device),
-        action=torch.tensor(np.array(actions)[-self._config.context_length:]).reshape(1, -1, 1).to(torch.float32).to(self.device),
+        action=torch.tensor(np.array(actions)[-self._config.context_length:]).reshape(1, -1, 1).to(torch.int64).to(self.device),
         timestep=torch.tensor(np.array(steps)).to(torch.int64).reshape(1, 1, 1).to(self.device),
       )
